+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.3.4] - 2025-12-21
 - refacoring & bug fix
+- remove dead code
 
 ## [0.3.2] - 2025-12-08
 - hot fix